First cleanup and rename to harmonization only
