--- neuroimaging/modules/harmonization_plotly/__init__.py
@@ -1,3 +1,0 @@
-from .harmonization_plotly import MultiqcModule
-
-__all__ = ["MultiqcModule"]
