--- neuroimaging/modules/metricsinroi_harmonized/__init__.py
@@ -1,3 +1,0 @@
-from .metricsinroi_harmonized import MultiqcModule
-
-__all__ = ["MultiqcModule"]
