+++ pyproject.toml
@@ -41,8 +41,6 @@ packages = [
     "neuroimaging.modules",
     "neuroimaging.modules.tractometry",
     "neuroimaging.modules.metricsinroi",
-    "neuroimaging.modules.metricsinroi_harmonized",
-    "neuroimaging.modules.harmonization_plotly",
     "neuroimaging.modules.harmonization",
     "neuroimaging.modules.cortical",
     "neuroimaging.modules.subcortical",
@@ -70,8 +68,7 @@ framewise_displacement = "neuroimaging.modules.framewise_displacement:MultiqcMod
 coverage = "neuroimaging.modules.coverage:MultiqcModule"
 streamline_count = "neuroimaging.modules.streamline_count:MultiqcModule"
 metricsinroi = "neuroimaging.modules.metricsinroi:MultiqcModule"
-metricsinroi_harmonized = "neuroimaging.modules.metricsinroi_harmonized:MultiqcModule"
-harmonization_plotly = "neuroimaging.modules.harmonization_plotly.harmonization_plotly:MultiqcModule"
+harmonization = "neuroimaging.modules.harmonization.harmonization:MultiqcModule"
 
 [project.entry-points."multiqc.cli_options.v1"]
 single_subject = "neuroimaging.cli:single_subject"
